Rename a class.
