+++ registration/src/test/java/org/spine3/samples/lobby/registration/seat/availability/SeatsAvailabilityAggregateShould.java
@@ -37,7 +37,7 @@ import java.util.Map;
 
 import static com.google.common.base.Throwables.propagate;
 import static org.junit.Assert.*;
-import static org.spine3.samples.lobby.common.util.CommonMessageFactory.*;
+import static org.spine3.samples.lobby.common.util.IdFactory.*;
 import static org.spine3.samples.lobby.registration.util.CollectionUtils.findById;
 import static org.spine3.samples.lobby.registration.util.MessageFactory.*;
 import static org.spine3.util.Identifiers.newUuid;
